Add shell object, decouple player health bar a bit
Shell object prevents player from leaving a given area even while they are noclipping.
Shells can be removed by collecting enough items (items not implemented yet).
The amount of items required for a shell to fade is set on the shell object.
There's a shell prefab as an example, and an example that appears in the Main scene.