Add trap movement
TrigTranslation moves the traps. Magnitude is how big its movement path is.
Frequency is how quickly it completes a wave along each axis.
Phase is its offset along the path for each axis.
The default settings will move a trap in a circular motion (if the magnitude is increased), and one can experiment with other values from there.